backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
plotly==5.18.0
matplotlib==3.8.2
openpyxl==3.1.2
python-multipart==0.0.6
reportlab==4.0.8
kaleido==0.2.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-generativeai==0.3.2
python-dotenv==1.0.0
